Adds tests for all character commands.

+++ src/Console/Command/Character/CharacterResetViewpointCommand.php
@@ -3,32 +3,31 @@ declare(strict_types=1);
 
 namespace LotGD\Core\Console\Command\Character;
 
-use LotGD\Core\Console\Command\BaseCommand;
+use Exception;
 use LotGD\Core\Models\Character;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputDefinition;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 /**
  * Resets the viewpoint of a given character.
  */
-class CharacterResetViewpointCommand extends BaseCommand
+class CharacterResetViewpointCommand extends CharacterBaseCommand
 {
     /**
      * @inheritDoc
      */
     protected function configure()
     {
-        $this->setName('character:resetViewpoint')
-            ->setDescription('Resets the viewpoint of a given character.')
+        $this->setName($this->namespaced("resetViewpoint"))
+            ->setDescription("Resets the viewpoint of a given character.")
             ->setDefinition(
                 new InputDefinition([
-                    new InputArgument("id", InputArgument::REQUIRED, "ID of the character"),
-                ])
+                    $this->getCharacterIdArgumentDefinition(),
+                ]),
             )
         ;
     }
@@ -39,6 +38,7 @@ class CharacterResetViewpointCommand extends BaseCommand
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $em = $this->game->getEntityManager();
+        $logger = $this->getCliLogger();
 
         $io = new SymfonyStyle($input, $output);
         $id = $input->getArgument("id");
@@ -51,11 +51,22 @@ class CharacterResetViewpointCommand extends BaseCommand
             return Command::FAILURE;
         }
 
-        $em->remove($character->getViewpoint());
-        $character->setViewpoint(null);
+        if ($character->getViewpoint() === null) {
+            $io->info("Character does not have a viewpoint yet.");
+        } else {
+            try {
+                $em->remove($character->getViewpoint());
+                $character->setViewpoint(null);
 
-        # Save
-        $em->flush();
+                $io->success("Viewpoint was successfully reset.");
+
+                # Save
+                $em->flush();
+            } catch (Exception $e) {
+                $io->error("Resetting the viewpoint was not possible. Reason: {$e}");
+                return Command::FAILURE;
+            }
+        }
 
         return Command::SUCCESS;
     }
